+++ packages/plugins/eslint-plugin-react-core/src/rules/no-class-component.ts
@@ -32,7 +32,7 @@ export default createRule<[], MessageID>({
   },
   name: RULE_NAME,
   create(context) {
-    const { ctx, listeners } = useComponentCollectorLegacy(context);
+    const { ctx, listeners } = useComponentCollectorLegacy();
 
     return {
       ...listeners,