+++ packages/plugins/eslint-plugin-react-core/src/rules/no-redundant-should-component-update.ts
@@ -33,7 +33,7 @@ export default createRule<[], MessageID>({
   },
   name: RULE_NAME,
   create(context) {
-    const { ctx, listeners } = useComponentCollectorLegacy(context);
+    const { ctx, listeners } = useComponentCollectorLegacy();
 
     return {
       ...listeners,