+++ packages/plugins/eslint-plugin-react-hooks-extra/src/rules/ensure-use-memo-has-non-empty-deps.ts
@@ -1,12 +1,12 @@
 import { is, NodeType } from "@eslint-react/ast";
 import { isReactHookCall, isReactHookCallWithNameLoose, isUseMemoCall } from "@eslint-react/core";
-import { getPragmaFromContext } from "@eslint-react/jsx";
-import { ESLintSettingsSchema, parseSchema } from "@eslint-react/shared";
+import { ESLintSettingsSchema } from "@eslint-react/shared";
 import { findVariable, getVariableInit } from "@eslint-react/var";
 import type { ESLintUtils } from "@typescript-eslint/utils";
 import { Function as F, Option as O } from "effect";
 import { type ConstantCase } from "string-ts";
 import { match } from "ts-pattern";
+import { parse } from "valibot";
 
 import { createRule } from "../utils";
 export const RULE_NAME = "ensure-use-memo-has-non-empty-deps";
@@ -27,14 +27,13 @@ export default createRule<[], MessageID>({
   },
   name: RULE_NAME,
   create(context) {
-    const alias = parseSchema(ESLintSettingsSchema, context.settings).reactOptions?.additionalHooks?.useMemo ?? [];
-    const pragma = getPragmaFromContext(context);
+    const alias = parse(ESLintSettingsSchema, context.settings).reactOptions?.additionalHooks?.useMemo ?? [];
 
     return {
       CallExpression(node) {
         const initialScope = context.sourceCode.getScope(node);
         if (!isReactHookCall(node)) return;
-        if (!isUseMemoCall(node, context, pragma) && !alias.some(F.flip(isReactHookCallWithNameLoose)(node))) return;
+        if (!isUseMemoCall(node, context) && !alias.some(F.flip(isReactHookCallWithNameLoose)(node))) return;
         const [_, deps] = node.arguments;
         if (!deps) {
           context.report({