+++ packages/utilities/jsx/src/element/index.ts
@@ -1,4 +1,2 @@
-export * from "./api";
 export * from "./element-type";
-export * from "./hierarchy";
-export * from "./is-element";
+export * from "./is";